Fix the asset uploading
We used to try to upload the binaries to a release defined in the current (sbt/sbtn-dist) repository, which failed because the releases are defined in the sbt/sbt repo.

We fix that by fetching the sbt/sbt release that matches the tag that triggered the build and we upload to that release.